@mui/icons-material library was added and changes were made to the Nav
