+++ components/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./Navbar";
 export * from "./CustonLink";
+export * from "./ItemsMenu";
